Implement live preview feature.
When the mouse cursor is hovered on a tab, a thumbnail preview of the web view associated with the tab is shown in a small pop-up. The preview is live, if the web page contains e.g. a video, then the video can be seen in the preview.  Right now updating is implemented by forcing a repaint, we need to find a better solution (likely should be done in QtWebKit).
Also scaling is done in software. An OpenGL widget is a another candidate. Or even using the full-page zoom feature (once it is landed in WebKit trunk).